# frontend/src/data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Fee Management API",
    "description": "RESTful API for school fee management system",
    "contact": {
      "name": "School Fee App Team",
      "email": "support@schoolfeeapp.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    },
    {
      "url": "http://localhost:8081",
      "description": "Docker Development"
    }
  ],
  "paths": {
...
    "/api/v1/results/terms/{termId}/unpublish": {
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "unpublishResults",
        "parameters": [
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublishResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/terms/{termId}/publish": {
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "publishResults",
        "parameters": [
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublishResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/scores/{scoreId}": {
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "updateScore",
        "parameters": [
          {
            "name": "scoreId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUpdateScoreResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/report-cards/{studentId}/term/{termId}/teacher-comment": {
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "addTeacherComment",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReportCommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/report-cards/{studentId}/term/{termId}/principal-comment": {
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "addPrincipalComment",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReportCommentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/grading-rules": {
      "get": {
        "tags": [
          "result-controller"
        ],
        "operationId": "getGradingRules",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGradingRuleResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "configureGradingRules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradingRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseGradingRuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/ca-config": {
      "get": {
        "tags": [
          "result-controller"
        ],
        "operationId": "getCaConfig",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCaConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "result-controller"
        ],
        "operationId": "configureCa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCaConfigResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/report-cards": {
      "post": {
        "tags": [
          "result-controller"
        ],
        "operationId": "generateReportCards",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportCardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReportCardJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/rankings/recompute": {
      "post": {
        "tags": [
          "result-controller"
        ],
        "operationId": "recomputeRankings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecomputeRankingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMapStringString"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/exam-scores": {
      "post": {
        "tags": [
          "result-controller"
        ],
        "operationId": "enterExamScores",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExamScoreResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/ca-scores": {
      "post": {
        "tags": [
          "result-controller"
        ],
        "operationId": "enterCaScores",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCaScoreResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/students/{studentId}/term/{termId}": {
      "get": {
        "tags": [
          "result-controller"
        ],
        "operationId": "getStudentResult",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseStudentResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/results/report-cards/jobs/{jobId}": {
      "get": {
        "tags": [
          "result-controller"
        ],
        "operationId": "getReportCardJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseReportCardJobResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/results/classes/{classId}/term/{termId}/result-sheet": {
      "get": {
        "tags": [
          "result-controller"
        ],
        "operationId": "getClassResultSheet",
        "parameters": [
          {
            "name": "classId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseClassResultSheetResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Meta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResponsePublishResultResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PublishResultResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PublishResultResponse": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publishedBy": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UpdateScoreRequest": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "score"
        ]
      },
      "ApiResponseUpdateScoreResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UpdateScoreResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateScoreResponse": {
        "type": "object",
        "properties": {
          "scoreId": {
            "type": "string",
            "format": "uuid"
          },
          "newScore": {
            "type": "number"
          },
          "previousScore": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CommentRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "comment"
        ]
      },
      "ApiResponseReportCommentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ReportCommentResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReportCommentResponse": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "comment": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GradingRuleRequest": {
        "type": "object",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/JsonNode"
          }
        },
        "required": [
          "config"
        ]
      },
      "JsonNode": {
        "type": "object"
      },
      "ApiResponseGradingRuleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/GradingRuleResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GradingRuleResponse": {
        "type": "object",
        "properties": {
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "gradesCount": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "CaComponentRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "maxScore": {
            "type": "integer",
            "format": "int32"
          },
          "weightPercentage": {
            "type": "number",
            "format": "double"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CaConfigRequest": {
        "type": "object",
        "properties": {
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CaComponentRequest"
            }
          },
          "examWeightPercentage": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "components"
        ]
      },
      "ApiResponseCaConfigResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CaConfigResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CaConfigResponse": {
        "type": "object",
        "properties": {
          "componentCount": {
            "type": "integer",
            "format": "int32"
          },
          "examWeightPercentage": {
            "type": "number",
            "format": "double"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseMapStringString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReportCardRequest": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "studentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "includeAttendance": {
            "type": "boolean"
          },
          "includeTeacherComment": {
            "type": "boolean"
          },
          "includePrincipalComment": {
            "type": "boolean"
          },
          "teacherComment": {
            "type": "string"
          },
          "format": {
            "type": "string"
          }
        },
        "required": [
          "classId",
          "studentIds",
          "termId"
        ]
      },
      "ApiResponseReportCardJobResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ReportCardJobResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReportCardJobResponse": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "totalStudents": {
            "type": "integer",
            "format": "int32"
          },
          "completedStudents": {
            "type": "integer",
            "format": "int32"
          },
          "failedStudents": {
            "type": "integer",
            "format": "int32"
          },
          "downloadUrl": {
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "RecomputeRankingsRequest": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "classId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "classId",
          "termId"
        ]
      },
      "ExamScoreRequest": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "string",
            "format": "uuid"
          },
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "maxScore": {
            "type": "integer",
            "format": "int32"
          },
          "scores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreEntry"
            }
          }
        },
        "required": [
          "classId",
          "examId",
          "scores",
          "subjectId",
          "termId"
        ]
      },
      "ScoreEntry": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "score",
          "studentId"
        ]
      },
      "ApiResponseExamScoreResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ExamScoreResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CaScoreRequest": {
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid"
          },
          "classId": {
            "type": "string",
            "format": "uuid"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "caComponentId": {
            "type": "string",
            "format": "uuid"
          },
          "maxScore": {
            "type": "integer",
            "format": "int32"
          },
          "scores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreEntry"
            }
          }
        },
        "required": [
          "caComponentId",
          "classId",
          "scores",
          "subjectId",
          "termId"
        ]
      },
      "ApiResponseCaScoreResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/CaScoreResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseStudentResultResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/StudentResultResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseClassResultSheetResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ClassResultSheetResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "errors": {},
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}